chore(webgl): .ts file for Program class
